+++ backend/src/main/java/com/mapbefine/mapbefine/image/application/ImageService.java
@@ -1,10 +1,10 @@
-package com.mapbefine.mapbefine.s3.application;
+package com.mapbefine.mapbefine.image.application;
 
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 @Service
-public interface S3Service {
+public interface ImageService {
 
     String upload(MultipartFile multipartFile);
