+++ backend/src/main/java/com/mapbefine/mapbefine/image/application/S3ImageService.java
@@ -1,7 +1,7 @@
-package com.mapbefine.mapbefine.s3.application;
+package com.mapbefine.mapbefine.image.application;
 
-import com.mapbefine.mapbefine.s3.domain.S3Client;
-import com.mapbefine.mapbefine.s3.domain.UploadFile;
+import com.mapbefine.mapbefine.image.domain.S3Client;
+import com.mapbefine.mapbefine.image.domain.UploadFile;
 import java.io.IOException;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Profile;
@@ -10,13 +10,13 @@ import org.springframework.web.multipart.MultipartFile;
 
 @Service
 @Profile("!test") // TODO : Test 시에는 TestS3ServiceImpl 를 Component 로 띄우기 위해 Profile 을 분리하였습니다.
-public class S3ServiceImpl implements S3Service {
+public class S3ImageService implements ImageService {
 
     @Value("${prefix.upload.path}")
     private String prefixUploadPath;
     private final S3Client s3Client;
 
-    public S3ServiceImpl(S3Client s3Client) {
+    public S3ImageService(S3Client s3Client) {
         this.s3Client = s3Client;
     }
