+++ backend/src/main/java/com/mapbefine/mapbefine/image/domain/UploadFile.java
@@ -1,4 +1,4 @@
-package com.mapbefine.mapbefine.s3.domain;
+package com.mapbefine.mapbefine.image.domain;
 
 import java.io.ByteArrayInputStream;
 import java.io.File;
