+++ backend/src/main/java/com/mapbefine/mapbefine/pin/application/PinCommandService.java
@@ -3,7 +3,7 @@ package com.mapbefine.mapbefine.pin.application;
 import static com.mapbefine.mapbefine.pin.exception.PinErrorCode.FORBIDDEN_PIN_CREATE_OR_UPDATE;
 import static com.mapbefine.mapbefine.pin.exception.PinErrorCode.ILLEGAL_PIN_ID;
 import static com.mapbefine.mapbefine.pin.exception.PinErrorCode.ILLEGAL_PIN_IMAGE_ID;
-import static com.mapbefine.mapbefine.s3.exception.S3ErrorCode.IMAGE_FILE_IS_NULL;
+import static com.mapbefine.mapbefine.image.exception.S3ErrorCode.IMAGE_FILE_IS_NULL;
 import static com.mapbefine.mapbefine.topic.exception.TopicErrorCode.ILLEGAL_TOPIC_ID;
 
 import com.mapbefine.mapbefine.auth.domain.AuthMember;
@@ -22,8 +22,8 @@ import com.mapbefine.mapbefine.pin.dto.request.PinImageCreateRequest;
 import com.mapbefine.mapbefine.pin.dto.request.PinUpdateRequest;
 import com.mapbefine.mapbefine.pin.exception.PinException.PinBadRequestException;
 import com.mapbefine.mapbefine.pin.exception.PinException.PinForbiddenException;
-import com.mapbefine.mapbefine.s3.application.S3Service;
-import com.mapbefine.mapbefine.s3.exception.S3Exception.S3BadRequestException;
+import com.mapbefine.mapbefine.image.application.ImageService;
+import com.mapbefine.mapbefine.image.exception.S3Exception.S3BadRequestException;
 import com.mapbefine.mapbefine.topic.domain.Topic;
 import com.mapbefine.mapbefine.topic.domain.TopicRepository;
 import com.mapbefine.mapbefine.topic.exception.TopicException.TopicBadRequestException;
@@ -45,7 +45,7 @@ public class PinCommandService {
     private final TopicRepository topicRepository;
     private final MemberRepository memberRepository;
     private final PinImageRepository pinImageRepository;
-    private final S3Service s3Service;
+    private final ImageService imageService;
 
     public PinCommandService(
             PinRepository pinRepository,
@@ -53,14 +53,14 @@ public class PinCommandService {
             TopicRepository topicRepository,
             MemberRepository memberRepository,
             PinImageRepository pinImageRepository,
-            S3Service s3Service
+            ImageService imageService
     ) {
         this.pinRepository = pinRepository;
         this.locationRepository = locationRepository;
         this.topicRepository = topicRepository;
         this.memberRepository = memberRepository;
         this.pinImageRepository = pinImageRepository;
-        this.s3Service = s3Service;
+        this.imageService = imageService;
     }
 
     public long save(
@@ -168,7 +168,7 @@ public class PinCommandService {
             throw new S3BadRequestException(IMAGE_FILE_IS_NULL);
         }
 
-        String imageUrl = s3Service.upload(image);
+        String imageUrl = imageService.upload(image);
         PinImage.createPinImageAssociatedWithPin(imageUrl, pin);
     }
